requirements.txt
deepgram-sdk
sounddevice
numpy
python-dotenv
fastapi
uvicorn
websockets
pypdf
groq